Fix up faceting and support prefixes for IRIs
