+++ ontogen/primitives/__init__.py
@@ -1,5 +1,7 @@
 from .datatypes import Datatype
-from .base import (Ontology, OwlEntity, OwlClass, OwlAnnotationProperty, OwlObjectProperty,
-                   OwlDataProperty, OwlThing,
-                   BASE_ENTITIES, COMMENT_ENTITY_NAME,
+from .base import (Ontology, OwlEntity, OwlAnnotationProperty,
+                   OwlDataProperty)
+from .vars import (BASE_ENTITIES, COMMENT_ENTITY_NAME,
                    LABEL_ENTITY_NAME, PROPERTY_ENTITIES)
+from .properties import OwlObjectProperty
+from .classes import OwlIndividual, OwlClass, OwlThing
